+++ VK-Challenge/VK-Challenge/support/Assets.xcassets/Placeholder/imgAvatarPlaceholder.imageset/Contents.json
@@ -2,7 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "ava-placeholder.png",
+      "filename" : "me.jpg",
       "scale" : "1x"
     },
     {
